restructures numerical solvers. re-implements all of the methods in pq_numerical_test.py as callable methods
