+++ pdaggerq/numerical/methods/cc3.py
@@ -1,7 +1,8 @@
 from pdaggerq.numerical.solvers.cc import cc 
 from pdaggerq.numerical.codegen.autogen import cc_residual
+from pdaggerq.numerical.codegen.autogen import cc3_triples_residual
 
-class CCSD:
+class CC3:
     def __init__(self, wfn, mol, **kwargs):
 
         self.wfn = wfn
@@ -18,7 +19,7 @@ class CCSD:
         local_namespace = {}
         
         # Generate equations
-        T = ['t1', 't2']
+        T = ['t1', 't2', 't3']
         
         cc_energy_func = cc_residual('cc_energy',
             T,
@@ -40,10 +41,17 @@ class CCSD:
             't2_residual',
             pq_graph_options = self.pq_graph_options
         )
+
+        t3_residual_func = cc3_triples_residual('r3',
+            [['e3(i,j,k,c,b,a)']],
+            't3_residual',
+            pq_graph_options = self.pq_graph_options
+        ) 
         
         # Execute the code strings in memory
         exec(t1_residual_func, globals(), local_namespace)
         exec(t2_residual_func, globals(), local_namespace)
+        exec(t3_residual_func, globals(), local_namespace)
 
         # amplitude dictionaries to pass into the solver
         t1 = {
@@ -56,7 +64,12 @@ class CCSD:
             'spins' : ['aaaa', 'abab', 'bbbb'],
             'residual' : local_namespace["t2_residual"]
         }
-        self.T_list = [t1, t2]
+        t3 = {
+            'spaces' : 'vvvooo',
+            'spins' : ['aaaaaa', 'aabaab', 'abbabb', 'bbbbbb'],
+            'residual' : local_namespace["t3_residual"]
+        }
+        self.T_list = [t1, t2, t3]
 
         self.cc_energy = {}
         exec(cc_energy_func, globals(), self.cc_energy)
@@ -75,63 +88,3 @@ class CCSD:
         en = self.cc_solver.t_solver()
 
         return en
-
-    def lambda_solver(self):
-
-        # Import pq codegen functions 
-        from pdaggerq.numerical.codegen.autogen import lambda_cc_residual
-        from pdaggerq.numerical.codegen.autogen import lambda_cc_pseudoenergy
-
-        # Create an empty dictionary to hold the pq-generated equations
-        local_namespace = {}
-
-        # Generate lambda equations
-        T = ['t1', 't2']
-        L = [['l1'], ['l2']]
-
-        cc_pseudoenergy_func = lambda_cc_pseudoenergy('cc_pseudoenergy',
-            L,
-            [['1']],
-            'cc_pseudoenergy',
-            pq_graph_options = self.pq_graph_options
-        )
-        l1_residual_func = lambda_cc_residual('r1',
-            T,
-            L,
-            'e1(a,i)',
-            'l1_residual',
-            pq_graph_options = self.pq_graph_options
-        )
-
-        l2_residual_func = lambda_cc_residual('r2',
-            T,
-            L,
-            'e2(a,b,j,i)',
-            'l2_residual',
-            pq_graph_options = self.pq_graph_options
-        )
-
-        # Execute the code strings in memory
-        exec(cc_pseudoenergy_func, globals(), local_namespace)
-        exec(l1_residual_func, globals(), local_namespace)
-        exec(l2_residual_func, globals(), local_namespace)
-
-        # lambda amplitude dictionaries to pass into the solver
-        l1 = {
-            'spaces' : 'vo',
-            'spins' : ['aa', 'bb'],
-            'residual' : local_namespace["l1_residual"]
-        }
-        l2 = {
-            'spaces' : 'vvoo',
-            'spins' : ['aaaa', 'abab', 'bbbb'],
-            'residual' : local_namespace["l2_residual"]
-        }
-
-        # initialize lambdas in cc_solver
-        self.cc_solver.initialize_lambda([l1, l2], cc_pseudoenergy_func = local_namespace['cc_pseudoenergy'])
-
-        en = self.cc_solver.lambda_solver()
-        return en
-
-
