+++ pdaggerq/numerical/methods/qed_ccsd_21.py
@@ -1,7 +1,7 @@
 from pdaggerq.numerical.solvers.cc import cc 
 from pdaggerq.numerical.codegen.autogen import cc_residual
 
-class CCSD:
+class QED_CCSD_21:
     def __init__(self, wfn, mol, **kwargs):
 
         self.wfn = wfn
@@ -13,37 +13,70 @@ class CCSD:
         self.e_convergence = kwargs.get('e_convergence', 1e-8)
         self.r_convergence = kwargs.get('r_convergence', 1e-6)
         self.max_iter = kwargs.get('max_iter', 50)
+        self.cavity_lambda = kwargs.get('cavity_lambda', 0.0)
+        self.cavity_frequency = kwargs.get('cavity_frequency', 0.07349864501573)
 
         # Create an empty dictionary to hold the pq-generated equations
         local_namespace = {}
-        
+       
+        T = ['t1', 't2', 't0,1', 't1,1', 't2,1']
+
         # Generate equations
-        T = ['t1', 't2']
-        
         cc_energy_func = cc_residual('cc_energy',
             T,
             [['1']],
             'cc_energy',
+            is_qed = True,
             pq_graph_options = self.pq_graph_options
         )
-        
+
         t1_residual_func = cc_residual('r1',
             T,
             [['e1(i,a)']],
             't1_residual',
+            is_qed = True,
             pq_graph_options = self.pq_graph_options
         )
-        
+
         t2_residual_func = cc_residual('r2',
             T,
             [['e2(i,j,b,a)']],
             't2_residual',
+            is_qed = True,
             pq_graph_options = self.pq_graph_options
         )
-        
+
+        t0_1p_residual_func = cc_residual('r0_1p',
+            T,
+            [['B-']],
+            't0_1p_residual',
+            is_qed = True,
+            pq_graph_options = self.pq_graph_options
+        )
+
+        t1_1p_residual_func = cc_residual('r1_1p',
+            T,
+            [['B-','e1(i,a)']],
+            't1_1p_residual',
+            is_qed = True,
+            pq_graph_options = self.pq_graph_options
+        )
+
+        t2_1p_residual_func = cc_residual('r2_1p',
+            T,
+            [['B-','e2(i,j,b,a)']],
+            't2_1p_residual',
+            is_qed = True,
+            pq_graph_options = self.pq_graph_options
+        )
+
         # Execute the code strings in memory
+        exec(cc_energy_func, globals(), local_namespace)
         exec(t1_residual_func, globals(), local_namespace)
         exec(t2_residual_func, globals(), local_namespace)
+        exec(t0_1p_residual_func, globals(), local_namespace)
+        exec(t1_1p_residual_func, globals(), local_namespace)
+        exec(t2_1p_residual_func, globals(), local_namespace)
 
         # amplitude dictionaries to pass into the solver
         t1 = {
@@ -56,7 +89,24 @@ class CCSD:
             'spins' : ['aaaa', 'abab', 'bbbb'],
             'residual' : local_namespace["t2_residual"]
         }
-        self.T_list = [t1, t2]
+        t0_1p = {
+            'nph' : 1,
+            'residual' : local_namespace["t0_1p_residual"]
+        }
+        t1_1p = {
+            'nph' : 1,
+            'spaces' : 'vo',
+            'spins' : ['aa', 'bb'],
+            'residual' : local_namespace["t1_1p_residual"]
+        }
+        t2_1p = {
+            'nph' : 1,
+            'spaces' : 'vvoo',
+            'spins' : ['aaaa', 'abab', 'bbbb'],
+            'residual' : local_namespace["t2_1p_residual"]
+        } 
+        
+        self.T_list = [t1, t2, t0_1p, t1_1p, t2_1p]
 
         self.cc_energy = {}
         exec(cc_energy_func, globals(), self.cc_energy)
@@ -69,69 +119,12 @@ class CCSD:
             self.mol,
             nfzc = self.nfzc,
             cc_energy_func = self.cc_energy["cc_energy"],
-            T_list = self.T_list
+            is_qed = True,
+            T_list = self.T_list,
+            cavity_lambda = self.cavity_lambda,
+            cavity_frequency= self.cavity_frequency
         )
         
         en = self.cc_solver.t_solver()
 
         return en
-
-    def lambda_solver(self):
-
-        # Import pq codegen functions 
-        from pdaggerq.numerical.codegen.autogen import lambda_cc_residual
-        from pdaggerq.numerical.codegen.autogen import lambda_cc_pseudoenergy
-
-        # Create an empty dictionary to hold the pq-generated equations
-        local_namespace = {}
-
-        # Generate lambda equations
-        T = ['t1', 't2']
-        L = [['l1'], ['l2']]
-
-        cc_pseudoenergy_func = lambda_cc_pseudoenergy('cc_pseudoenergy',
-            L,
-            [['1']],
-            'cc_pseudoenergy',
-            pq_graph_options = self.pq_graph_options
-        )
-        l1_residual_func = lambda_cc_residual('r1',
-            T,
-            L,
-            'e1(a,i)',
-            'l1_residual',
-            pq_graph_options = self.pq_graph_options
-        )
-
-        l2_residual_func = lambda_cc_residual('r2',
-            T,
-            L,
-            'e2(a,b,j,i)',
-            'l2_residual',
-            pq_graph_options = self.pq_graph_options
-        )
-
-        # Execute the code strings in memory
-        exec(cc_pseudoenergy_func, globals(), local_namespace)
-        exec(l1_residual_func, globals(), local_namespace)
-        exec(l2_residual_func, globals(), local_namespace)
-
-        # lambda amplitude dictionaries to pass into the solver
-        l1 = {
-            'spaces' : 'vo',
-            'spins' : ['aa', 'bb'],
-            'residual' : local_namespace["l1_residual"]
-        }
-        l2 = {
-            'spaces' : 'vvoo',
-            'spins' : ['aaaa', 'abab', 'bbbb'],
-            'residual' : local_namespace["l2_residual"]
-        }
-
-        # initialize lambdas in cc_solver
-        self.cc_solver.initialize_lambda([l1, l2], cc_pseudoenergy_func = local_namespace['cc_pseudoenergy'])
-
-        en = self.cc_solver.lambda_solver()
-        return en
-
-
