+++ pdaggerq/numerical/solvers/eomcc.py
@@ -16,7 +16,7 @@
 #   limitations under the License.
 
 """
-eom-ccsd sigma vectors and the full eom-ccsd hamiltonian
+eom-cc solver
 """
 
 import numpy as np
@@ -31,10 +31,10 @@ import types
 
 import copy
 
-class eom_ccsd:
+class eomcc:
 
     def __init__(self, 
-                 ccsd,
+                 cc,
                  nstates = 5,
                  R_list=[],
                  L_list=[],
@@ -42,26 +42,29 @@ class eom_ccsd:
         """
         initialize EOMCC class
 
-        :params ccsd: the ccsd class
+        :params cc: the cc class
         :params R_list: list of R operator dictionaries
         :params L_list: list of L operator dictionaries
         """
 
-        self.ccsd = ccsd
+        self.cc = cc
         self.nstates = nstates
         self.R_list = R_list
         self.L_list = L_list
 
         if density_matrix_func is not None:
             self.density_matrix = types.MethodType(density_matrix_func, self)
-
-    def right_solver(self):
+        
+    def right_solver(self, R_list = None):
 
         print('    ==> right-hand EOMCC <==')
         print('')
+       
+        if R_list is not None: 
+            self.R_list = R_list
 
         # build Hbar operator object
-        Hbar = HbarOperator(self.ccsd, R_list = self.R_list)
+        Hbar = HbarOperator(self.cc, R_list = self.R_list)
 
         # diagonalize Hbar
         dim = Hbar.right_amplitude_size
@@ -74,27 +77,30 @@ class eom_ccsd:
 
         print('    %5s %20s %20s' % ('state', 'total energy','excitation energy'))
         for i in range (self.nstates):
-            print('    %5i %20.12f %20.12f' % ( i, ex[i].real + self.ccsd.energy + self.ccsd.efzc + self.ccsd.nuclear_repulsion_energy, ex[i].real))
+            print('    %5i %20.12f %20.12f' % ( i, ex[i].real + self.cc.energy + self.cc.efzc + self.cc.nuclear_repulsion_energy, ex[i].real))
 
         print('')
        
         # save energies and right-hand amplitudes
-        self.eom_cc_energy = []
+        self.eomcc_energy = []
         for i in range (self.nstates):
-            self.eom_cc_energy.append(ex[i])
+            self.eomcc_energy.append(ex[i])
 
         self.R = []
         for i in range (self.nstates):
             Hbar.unpack_eom_vectors(rvec[:, i], Hbar.R, Hbar.R_meta)
             self.R.append(copy.deepcopy(Hbar.R))
 
-    def left_solver(self):
+    def left_solver(self, L_list = None):
 
         print('    ==> left-hand EOMCC <==')
         print('')
+        
+        if L_list is not None:
+            self.L_list = L_list
 
         # build Hbar operator object
-        Hbar = HbarOperator(self.ccsd, L_list = self.L_list)
+        Hbar = HbarOperator(self.cc, L_list = self.L_list)
 
         # diagonalize Hbar
         dim = Hbar.left_amplitude_size
@@ -107,24 +113,29 @@ class eom_ccsd:
 
         print('    %5s %20s %20s' % ('state', 'total energy','excitation energy'))
         for i in range (self.nstates):
-            print('    %5i %20.12f %20.12f' % ( i, ex[i].real + self.ccsd.energy + self.ccsd.efzc + self.ccsd.nuclear_repulsion_energy, ex[i].real))
+            print('    %5i %20.12f %20.12f' % ( i, ex[i].real + self.cc.energy + self.cc.efzc + self.cc.nuclear_repulsion_energy, ex[i].real))
 
         print('')
        
         # save energies and left-hand amplitudes
-        self.eom_cc_energy = []
+        self.eomcc_energy = []
         for i in range (self.nstates):
-            self.eom_cc_energy.append(ex[i])
+            self.eomcc_energy.append(ex[i])
 
         self.L = []
         for i in range (self.nstates):
             Hbar.unpack_eom_vectors(lvec[:, i], Hbar.L, Hbar.L_meta)
             self.L.append(copy.deepcopy(Hbar.L))
 
-    def oscillator_strengths(self):
+    def oscillator_strengths(self, density_matrix_func = None):
+
+        if density_matrix_func is not None:
+            self.density_matrix = types.MethodType(density_matrix_func, self)
+        else:
+            raise Exception("provide a density_matrix_func")
 
         # Pack eigenvectors for biorthogonalization
-        Hbar = HbarOperator(self.ccsd, L_list = self.L_list, R_list = self.R_list)
+        Hbar = HbarOperator(self.cc, L_list = self.L_list, R_list = self.R_list)
 
         dim = Hbar.left_amplitude_size
         R_mat = np.zeros((dim, self.nstates), dtype = np.complex128)
@@ -151,7 +162,7 @@ class eom_ccsd:
         print('    %7s %7s %10s %10s %10s %10s %10s %10s %10s' % ('state 1', 'state 2', 'x', 'y', 'z', 'x', 'y', 'z', 'osc'))
 
         from pdaggerq.numerical.utils.integrals import get_dipole_integrals_with_spin
-        dipole_aa, dipole_bb = get_dipole_integrals_with_spin(self.ccsd.wfn, nfzc = self.ccsd.nfzc)
+        dipole_aa, dipole_bb = get_dipole_integrals_with_spin(self.cc.wfn, nfzc = self.cc.nfzc)
 
         f = np.zeros((self.nstates, self.nstates), dtype=np.complex128)
 
@@ -163,31 +174,31 @@ class eom_ccsd:
                 tdp_ij = np.zeros((3), dtype=np.complex128)
 
                 for xyz in range (3):
-                    tdp_ij[xyz] += np.einsum('ij,ij->', tdm['aa_oo'], dipole_aa[xyz][self.ccsd.oa, self.ccsd.oa])
-                    tdp_ij[xyz] += np.einsum('ia,ia->', tdm['aa_ov'], dipole_aa[xyz][self.ccsd.oa, self.ccsd.va])
-                    tdp_ij[xyz] += np.einsum('ai,ai->', tdm['aa_vo'], dipole_aa[xyz][self.ccsd.va, self.ccsd.oa])
-                    tdp_ij[xyz] += np.einsum('ab,ab->', tdm['aa_vv'], dipole_aa[xyz][self.ccsd.va, self.ccsd.va])
-                    tdp_ij[xyz] += np.einsum('ij,ij->', tdm['bb_oo'], dipole_bb[xyz][self.ccsd.ob, self.ccsd.ob])
-                    tdp_ij[xyz] += np.einsum('ia,ia->', tdm['bb_ov'], dipole_bb[xyz][self.ccsd.ob, self.ccsd.vb])
-                    tdp_ij[xyz] += np.einsum('ai,ai->', tdm['bb_vo'], dipole_bb[xyz][self.ccsd.vb, self.ccsd.ob])
-                    tdp_ij[xyz] += np.einsum('ab,ab->', tdm['bb_vv'], dipole_bb[xyz][self.ccsd.vb, self.ccsd.vb])
+                    tdp_ij[xyz] += np.einsum('ij,ij->', tdm['aa_oo'], dipole_aa[xyz][self.cc.oa, self.cc.oa])
+                    tdp_ij[xyz] += np.einsum('ia,ia->', tdm['aa_ov'], dipole_aa[xyz][self.cc.oa, self.cc.va])
+                    tdp_ij[xyz] += np.einsum('ai,ai->', tdm['aa_vo'], dipole_aa[xyz][self.cc.va, self.cc.oa])
+                    tdp_ij[xyz] += np.einsum('ab,ab->', tdm['aa_vv'], dipole_aa[xyz][self.cc.va, self.cc.va])
+                    tdp_ij[xyz] += np.einsum('ij,ij->', tdm['bb_oo'], dipole_bb[xyz][self.cc.ob, self.cc.ob])
+                    tdp_ij[xyz] += np.einsum('ia,ia->', tdm['bb_ov'], dipole_bb[xyz][self.cc.ob, self.cc.vb])
+                    tdp_ij[xyz] += np.einsum('ai,ai->', tdm['bb_vo'], dipole_bb[xyz][self.cc.vb, self.cc.ob])
+                    tdp_ij[xyz] += np.einsum('ab,ab->', tdm['bb_vv'], dipole_bb[xyz][self.cc.vb, self.cc.vb])
 
                 tdm = self.density_matrix(j, i)
                     
                 tdp_ji = np.zeros((3), dtype=np.complex128)
 
                 for xyz in range (3):
-                    tdp_ji[xyz] += np.einsum('ij,ij->', tdm['aa_oo'], dipole_aa[xyz][self.ccsd.oa, self.ccsd.oa])
-                    tdp_ji[xyz] += np.einsum('ia,ia->', tdm['aa_ov'], dipole_aa[xyz][self.ccsd.oa, self.ccsd.va])
-                    tdp_ji[xyz] += np.einsum('ai,ai->', tdm['aa_vo'], dipole_aa[xyz][self.ccsd.va, self.ccsd.oa])
-                    tdp_ji[xyz] += np.einsum('ab,ab->', tdm['aa_vv'], dipole_aa[xyz][self.ccsd.va, self.ccsd.va])
-                    tdp_ji[xyz] += np.einsum('ij,ij->', tdm['bb_oo'], dipole_bb[xyz][self.ccsd.ob, self.ccsd.ob])
-                    tdp_ji[xyz] += np.einsum('ia,ia->', tdm['bb_ov'], dipole_bb[xyz][self.ccsd.ob, self.ccsd.vb])
-                    tdp_ji[xyz] += np.einsum('ai,ai->', tdm['bb_vo'], dipole_bb[xyz][self.ccsd.vb, self.ccsd.ob])
-                    tdp_ji[xyz] += np.einsum('ab,ab->', tdm['bb_vv'], dipole_bb[xyz][self.ccsd.vb, self.ccsd.vb])
+                    tdp_ji[xyz] += np.einsum('ij,ij->', tdm['aa_oo'], dipole_aa[xyz][self.cc.oa, self.cc.oa])
+                    tdp_ji[xyz] += np.einsum('ia,ia->', tdm['aa_ov'], dipole_aa[xyz][self.cc.oa, self.cc.va])
+                    tdp_ji[xyz] += np.einsum('ai,ai->', tdm['aa_vo'], dipole_aa[xyz][self.cc.va, self.cc.oa])
+                    tdp_ji[xyz] += np.einsum('ab,ab->', tdm['aa_vv'], dipole_aa[xyz][self.cc.va, self.cc.va])
+                    tdp_ji[xyz] += np.einsum('ij,ij->', tdm['bb_oo'], dipole_bb[xyz][self.cc.ob, self.cc.ob])
+                    tdp_ji[xyz] += np.einsum('ia,ia->', tdm['bb_ov'], dipole_bb[xyz][self.cc.ob, self.cc.vb])
+                    tdp_ji[xyz] += np.einsum('ai,ai->', tdm['bb_vo'], dipole_bb[xyz][self.cc.vb, self.cc.ob])
+                    tdp_ji[xyz] += np.einsum('ab,ab->', tdm['bb_vv'], dipole_bb[xyz][self.cc.vb, self.cc.vb])
 
-                en_i = self.eom_cc_energy[i]
-                en_j = self.eom_cc_energy[j]
+                en_i = self.eomcc_energy[i]
+                en_j = self.eomcc_energy[j]
                 w = en_j - en_i
 
                 f[i, j] = 2./3. * w * np.einsum('i,i->', tdp_ij, tdp_ji)
